# openapi.yaml
openapi: 3.0.3
info:
  title: OMM API
  version: 0.1.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        avatar_url:
          type: string
...
    RedeemVoucherRequest:
      type: object
      required: [voucher_code, employee_code]
      properties:
        voucher_code:
          type: string
        employee_code:
          type: string
    RedeemVoucherResponse:
      type: object
      properties:
        voucher_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        redeemed_at:
          type: string
          format: date-time
        voucher_code:
          type: string
...
paths:
...
  /users/profile:
    post:
      summary: Update profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /vouchers/redeem:
    post:
      summary: Redeem voucher
      operationId: redeemVoucher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemVoucherRequest"
      responses:
        "200":
          description: Voucher redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemVoucherResponse"
...